+++ src/chat/web/components/chat-container/styles.ts
@@ -1,8 +1,6 @@
 import { css } from "lit";
-import { themeColors } from "../../styles/theme-colors";
 
 export const chatContainerStyles = css`
-    ${themeColors}
     :host {
         display: flex;
         flex-direction: column;
@@ -40,11 +38,15 @@ export const chatContainerStyles = css`
         z-index: 10;
     }
 
+    collama-chat-modal {
+        flex: 0 0 auto;
+        display: flex;
+    }
+
     collama-chatinput {
         flex: 0 0 auto;
         display: flex;
         flex-direction: column;
         overflow: visible;
-        padding: 8px;
     }
 `;